[Refactor] WidgetManager 별도의 파일로 분리 #45
